+++ lib/internal/test_runner/mock/loader.js
@@ -1,77 +1,24 @@
 'use strict';
 const {
-  AtomicsNotify,
-  AtomicsStore,
   JSONStringify,
   SafeMap,
 } = primordials;
-const {
-  kBadExportsMessage,
-  kMockSearchParam,
-  kMockSuccess,
-  kMockExists,
-  kMockUnknownMessage,
-} = require('internal/test_runner/mock/mock');
+
+const kMockSearchParam = 'node-test-mock';
+const kBadExportsMessage = 'Cannot create mock because named exports ' +
+  'cannot be applied to the provided default export.';
+
 const { URL, URLParse } = require('internal/url');
 let debug = require('internal/util/debuglog').debuglog('test_runner', (fn) => {
   debug = fn;
 });
 
-// TODO(cjihrig): The mocks need to be thread aware because the exports are
-// evaluated on the thread that creates the mock. Before marking this API as
-// stable, one of the following issues needs to be implemented:
-// https://github.com/nodejs/node/issues/49472
-// or https://github.com/nodejs/node/issues/52219
-
 const mocks = new SafeMap();
 
-async function initialize(data) {
-  data?.port.on('message', ({ type, payload }) => {
-    debug('mock loader received message type "%s" with payload %o', type, payload);
-
-    if (type === 'node:test:register') {
-      const { baseURL } = payload;
-      const mock = mocks.get(baseURL);
-
-      if (mock?.active) {
-        debug('already mocking "%s"', baseURL);
-        sendAck(payload.ack, kMockExists);
-        return;
-      }
-
-      const localVersion = mock?.localVersion ?? 0;
-
-      debug('new mock version %d for "%s"', localVersion, baseURL);
-      mocks.set(baseURL, {
-        __proto__: null,
-        active: true,
-        cache: payload.cache,
-        exportNames: payload.exportNames,
-        format: payload.format,
-        hasDefaultExport: payload.hasDefaultExport,
-        localVersion,
-        url: baseURL,
-      });
-      sendAck(payload.ack);
-    } else if (type === 'node:test:unregister') {
-      const mock = mocks.get(payload.baseURL);
-
-      if (mock !== undefined) {
-        mock.active = false;
-        mock.localVersion++;
-      }
-
-      sendAck(payload.ack);
-    } else {
-      sendAck(payload.ack, kMockUnknownMessage);
-    }
-  });
-}
-
-async function resolve(specifier, context, nextResolve) {
+function resolve(specifier, context, nextResolve) {
   debug('resolve hook entry, specifier = "%s", context = %o', specifier, context);
 
-  const nextResolveResult = await nextResolve(specifier, context);
+  const nextResolveResult = nextResolve(specifier, context);
   const mockSpecifier = nextResolveResult.url;
 
   const mock = mocks.get(mockSpecifier);
@@ -95,7 +42,7 @@ async function resolve(specifier, context, nextResolve) {
   return { __proto__: null, url: href, format: nextResolveResult.format };
 }
 
-async function load(url, context, nextLoad) {
+function load(url, context, nextLoad) {
   debug('load hook entry, url = "%s", context = %o', url, context);
   const parsedURL = URLParse(url);
   if (parsedURL) {
@@ -105,7 +52,7 @@ async function load(url, context, nextLoad) {
   const baseURL = parsedURL ? parsedURL.href : url;
   const mock = mocks.get(baseURL);
 
-  const original = await nextLoad(url, context);
+  const original = nextLoad(url, context);
   debug('load hook, mock = %o', mock);
   if (mock?.active !== true) {
     return original;
@@ -130,14 +77,14 @@ async function load(url, context, nextLoad) {
     __proto__: null,
     format,
     shortCircuit: true,
-    source: await createSourceFromMock(mock, format),
+    source: createSourceFromMock(mock, format),
   };
 
   debug('load hook finished, result = %o', result);
   return result;
 }
 
-async function createSourceFromMock(mock, format) {
+function createSourceFromMock(mock, format) {
   // Create mock implementation from provided exports.
   const { exportNames, hasDefaultExport, url } = mock;
   const useESM = format === 'module' || format === 'module-typescript';
@@ -196,9 +143,12 @@ if (module.exports === null || typeof module.exports !== 'object') {
   return source;
 }
 
-function sendAck(buf, status = kMockSuccess) {
-  AtomicsStore(buf, 0, status);
-  AtomicsNotify(buf, 0);
-}
-
-module.exports = { initialize, load, resolve };
+module.exports = {
+  hooks: { __proto__: null, load, resolve },
+  mocks,
+  constants: {
+    __proto__: null,
+    kBadExportsMessage,
+    kMockSearchParam,
+  },
+};
