+++ lib/internal/test_runner/mock/mock.js
@@ -2,12 +2,9 @@
 const {
   ArrayPrototypePush,
   ArrayPrototypeSlice,
-  AtomicsStore,
-  AtomicsWait,
   Error,
   FunctionPrototypeBind,
   FunctionPrototypeCall,
-  Int32Array,
   ObjectDefineProperty,
   ObjectGetOwnPropertyDescriptor,
   ObjectGetPrototypeOf,
@@ -19,9 +16,6 @@ const {
   SafeMap,
   StringPrototypeSlice,
   StringPrototypeStartsWith,
-  globalThis: {
-    SharedArrayBuffer,
-  },
 } = primordials;
 const {
   codes: {
@@ -54,19 +48,10 @@ const {
   validateOneOf,
 } = require('internal/validators');
 const { MockTimers } = require('internal/test_runner/mock/mock_timers');
-const { strictEqual, notStrictEqual } = require('assert');
 const { Module } = require('internal/modules/cjs/loader');
-const { MessageChannel } = require('worker_threads');
 const { _load, _nodeModulePaths, _resolveFilename, isBuiltin } = Module;
 function kDefaultFunction() {}
 const enableModuleMocking = getOptionValue('--experimental-test-module-mocks');
-const kMockSearchParam = 'node-test-mock';
-const kMockSuccess = 1;
-const kMockExists = 2;
-const kMockUnknownMessage = 3;
-const kWaitTimeout = 5_000;
-const kBadExportsMessage = 'Cannot create mock because named exports ' +
-  'cannot be applied to the provided default export.';
 const kSupportedFormats = [
   'builtin',
   'commonjs-typescript',
@@ -76,6 +61,11 @@ const kSupportedFormats = [
   'module',
 ];
 let sharedModuleState;
+const {
+  hooks: mockHooks,
+  mocks,
+  constants: { kBadExportsMessage, kMockSearchParam },
+} = require('internal/test_runner/mock/loader');
 
 class MockFunctionContext {
   #calls;
@@ -201,8 +191,8 @@ class MockModuleContext {
     hasDefaultExport,
     namedExports,
     sharedState,
+    specifier,
   }) {
-    const ack = new Int32Array(new SharedArrayBuffer(4));
     const config = {
       __proto__: null,
       cache,
@@ -218,7 +208,6 @@ class MockModuleContext {
     this.#sharedState = sharedState;
     this.#restore = {
       __proto__: null,
-      ack,
       baseURL,
       cached: fullPath in Module._cache,
       format,
@@ -226,20 +215,29 @@ class MockModuleContext {
       value: Module._cache[fullPath],
     };
 
-    sharedState.loaderPort.postMessage({
-      __proto__: null,
-      type: 'node:test:register',
-      payload: {
+    const mock = mocks.get(baseURL);
+
+    if (mock?.active) {
+      debug('already mocking "%s"', baseURL);
+      throw new ERR_INVALID_STATE(
+        `Cannot mock '${specifier}'. The module is already mocked.`,
+      );
+    } else {
+      const localVersion = mock?.localVersion ?? 0;
+
+      debug('new mock version %d for "%s"', localVersion, baseURL);
+      mocks.set(baseURL, {
         __proto__: null,
-        ack,
-        baseURL,
+        url: baseURL,
         cache,
         exportNames: ObjectKeys(namedExports),
         hasDefaultExport,
         format,
-      },
-    });
-    waitForAck(ack);
+        localVersion,
+        active: true,
+      });
+    }
+
     delete Module._cache[fullPath];
     sharedState.mockExports.set(baseURL, {
       __proto__: null,
@@ -261,17 +259,12 @@ class MockModuleContext {
       Module._cache[this.#restore.fullPath] = this.#restore.value;
     }
 
-    AtomicsStore(this.#restore.ack, 0, 0);
-    this.#sharedState.loaderPort.postMessage({
-      __proto__: null,
-      type: 'node:test:unregister',
-      payload: {
-        __proto__: null,
-        ack: this.#restore.ack,
-        baseURL: this.#restore.baseURL,
-      },
-    });
-    waitForAck(this.#restore.ack);
+    const mock = mocks.get(this.#restore.baseURL);
+
+    if (mock !== undefined) {
+      mock.active = false;
+      mock.localVersion++;
+    }
 
     this.#sharedState.mockMap.delete(this.#restore.baseURL);
     this.#sharedState.mockMap.delete(this.#restore.fullPath);
@@ -654,7 +647,7 @@ class MockTracker {
     const hasFileProtocol = StringPrototypeStartsWith(filename, 'file://');
     const caller = hasFileProtocol ? filename : pathToFileURL(filename).href;
     const { format, url } = sharedState.moduleLoader.resolveSync(
-      mockSpecifier, caller, null,
+      mockSpecifier, caller, kEmptyObject,
     );
     debug('module mock, url = "%s", format = "%s", caller = "%s"', url, format, caller);
     if (format) { // Format is not yet known for ambiguous files when detection is enabled.
@@ -828,20 +821,13 @@ function setupSharedModuleState() {
   if (sharedModuleState === undefined) {
     const { mock } = require('test');
     const mockExports = new SafeMap();
-    const { port1, port2 } = new MessageChannel();
+    const { registerHooks } = require('internal/modules/customization_hooks');
     const moduleLoader = esmLoader.getOrInitializeCascadedLoader();
 
-    moduleLoader.register(
-      'internal/test_runner/mock/loader',
-      'node:',
-      { __proto__: null, port: port2 },
-      [port2],
-      true,
-    );
+    registerHooks(mockHooks);
 
     sharedModuleState = {
       __proto__: null,
-      loaderPort: port1,
       mockExports,
       mockMap: new SafeMap(),
       moduleLoader,
@@ -941,13 +927,6 @@ function findMethodOnPrototypeChain(instance, methodName) {
   return descriptor;
 }
 
-function waitForAck(buf) {
-  const result = AtomicsWait(buf, 0, 0, kWaitTimeout);
-
-  notStrictEqual(result, 'timed-out', 'test mocking synchronization failed');
-  strictEqual(buf[0], kMockSuccess);
-}
-
 function ensureNodeScheme(specifier) {
   if (!StringPrototypeStartsWith(specifier, 'node:')) {
     return `node:${specifier}`;
@@ -962,10 +941,5 @@ if (!enableModuleMocking) {
 
 module.exports = {
   ensureNodeScheme,
-  kBadExportsMessage,
-  kMockSearchParam,
-  kMockSuccess,
-  kMockExists,
-  kMockUnknownMessage,
   MockTracker,
 };
